small scale experiment on BERT
TASK:

Practical experiment

Note: for this option, knowledge of programming with Python is required.

For the practical experiment, the objective is to set up a small-scale experiment that investigates the implicit syntactic abilities of current neural network architectures. More specifically, we will investigate a neural transformer model known as BERT [1]. BERT is trained using a so-called masked language modeling objective: certain words of the original sentence are replaced by a special mask symbol, and the model's goal is to predict the original words.

The model only takes a list of words as input, without any explicit syntactic structure. Nevertheless, the model achieves good results on a wide range of natural language processing tasks. This leads us to the question whether the model maybe implicitly develops a certain notion of syntax. We will investigate this question by examining the model's predictions for sentences where syntactic structure plays an important role [2,3]. Consider the following examples:



The woman who bakes cookies for the children is ill.
The doctor who examines the patients is over there.

In both sentences, a proper subject-verb agreement requires an understanding of the hierarchical structure of the sentence, such that the actual subject (woman, doctor) is considered for the agreement instead of the attractor noun (children, patients). In our experiment, we will mask the main verb (is), and consider the model's predictions for both the singular (is) and the plural form (are). If the model is able to predict the correct form over the incorrect form - despite the presence of the attractor noun - then this is evidence that the model has implicitly learned a certain notion of syntax instead of mere sequential co-occurrence.

For this experiment, you will design a small-scale evaluation set of 50 sentences for your native language, similar to the example sentences above. If your native language does not have subject-verb agreement, you may carry out the experiment for English. You will then perform model predictions according to the setup described above, and describe and analyze the results in a project report (circa 5 pages).


Note that the original BERT model for English, as well as a multilingual version and a number of language-specific models for other languages, are available through the Python module transformers.